+++ core/metrics/base.go
@@ -2,9 +2,6 @@ package metrics
 
 import (
 	"context"
-	"go.opentelemetry.io/otel/sdk/metric"
-	"go.opentelemetry.io/otel/sdk/metric/aggregation"
-	"go.opentelemetry.io/otel/sdk/metric/metricdata"
 	"net/http"
 	"os"
 	"strconv"
@@ -142,37 +139,3 @@ func newBaseHandlerWithTracerProvider(buildInfo config.BuildInfo, tracerProvider
 }
 
 var _ Handler = &baseHandler{}
-
-// noOpExporter is a no-op metric exporter that prevents any metrics from being exported.
-type noOpMetricExporter struct{}
-
-func newNoOpExporter() noOpMetricExporter {
-	return noOpMetricExporter{}
-}
-
-// Temporality returns the temporality given instrument kind.
-func (n noOpMetricExporter) Temporality(kind metric.InstrumentKind) metricdata.Temporality {
-	return metric.DefaultTemporalitySelector(kind)
-}
-
-// Aggregation returns the aggregation for the given instrument kind.
-func (n noOpMetricExporter) Aggregation(kind metric.InstrumentKind) aggregation.Aggregation {
-	return metric.DefaultAggregationSelector(kind)
-}
-
-// Export exporter (no-op)
-func (n noOpMetricExporter) Export(ctx context.Context, metrics *metricdata.ResourceMetrics) error {
-	return nil
-}
-
-// ForceFlush exporter (no-op)
-func (n noOpMetricExporter) ForceFlush(ctx context.Context) error {
-	return nil
-}
-
-// Shutdown exporter (no-op)
-func (n noOpMetricExporter) Shutdown(ctx context.Context) error {
-	return nil
-}
-
-var _ metric.Exporter = &noOpMetricExporter{}
